Update to TDLib 1.8.1

+++ telegram/parser.go
@@ -42,7 +42,7 @@ func commandArgument(msgText string) string {
 
 func getsenderId(sender tdlib.MessageSender) int64 {
 	if sender.GetMessageSenderEnum() == "messageSenderUser" {
-		return int64(sender.(*tdlib.MessageSenderUser).UserId)
+		return sender.(*tdlib.MessageSenderUser).UserId
 	}
 	return sender.(*tdlib.MessageSenderChat).ChatId
 }
